Move basic properties to ResourceEntity
